% Auto-generated: do not edit by hand
\name{dashModalable}

\alias{dashModalable}

\title{DashModalable component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashModalable(children=NULL, id=NULL, modalChildren=NULL,
modalZoomable=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number | unnamed list. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{modalChildren}{A list of or a singular dash component, string or number | unnamed list. }

\item{modalZoomable}{Logical. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

